Warden Naval Focus Icons
Added focus icons that were already in game files with a shine animation. These images are somewhat relevant to their focus (or I tried to get them to be). I did not add new ones from online with a shine effect as it adds extra code clutter that will probably be removed later with the actual focus icons.